--- src/components/Dashboard/index.ts
@@ -1,12 +1,0 @@
-export { default as ChartSection } from './ChartSection';
-export { default as FAQSection } from './FAQSection';
-export { default as LineChartSection } from './LineChartSection';
-export { default as NoAccessLogoutModal } from './NoAccessLogoutModal';
-export { default as ProcessSection } from './ProcessSection';
-export { default as ProcessStep } from './ProcessStep';
-export { default as ScholarshipDurationSection } from './ScholarshipDurationSection';
-export { default as ScholarshipStudentTypeSection } from './ScholarshipStudentTypeSection';
-
-export * from './CompareOptionButtonSection';
-export * from './CompetencyAnalyzeChartSection';
-export * from './RadarChartSection';
