+++ src/hooks/queries/index.ts
@@ -1,10 +1,8 @@
 export { default as useDeleteSubmittedMileageMutation } from './useDeleteSubmittedMileageMutation';
-export { default as useGetCapabilityQuery } from './useGetCapabilityQuery';
 export { default as useGetEtcMileageQuery } from './useGetEtcMileageQuery';
 export { default as useGetIsAppliedScholarshipQuery } from './useGetIsAppliedScholarshipQuery';
 export { default as useGetMileageQuery } from './useGetMileageQuery';
 export { default as useGetScholarshipDurationQuery } from './useGetScholarshipDurationQuery';
-export { default as useGetSemesterCapabilityQuery } from './useGetSemesterCapabilityQuery';
 export { default as useGetSubmittedFileQuery } from './useGetSubmittedFileQuery';
 export { default as useGetSubmittedMileageQuery } from './useGetSubmittedMileageQuery';
 export { default as usePatchSubmittedMileageMutation } from './usePatchSubmittedMileageMutation';
