+++ src/main.tsx
@@ -1,4 +1,4 @@
-import '@/service/sentry/sentry';
+// import '@/service/sentry/sentry';
 import { StrictMode } from 'react';
 import { createRoot } from 'react-dom/client';
 import App from './App.tsx';
