+++ src/pages/DashboardPage/components/CompareOptionButtonSection.tsx
@@ -1,11 +1,6 @@
 import { Flex, ToggleButton } from '@/components';
 
-const compareOptionLabel = {
-  term: '동일 학기',
-  entryYear: '동일 학번',
-  major1: '동일 1 전공',
-  major2: '동일 2 전공',
-} as const;
+import { compareOptionLabel } from '../constants/compareOptionLabel';
 
 export const CompareOptionButtonSection = ({
   compareOption,
