+++ src/pages/DashboardPage/components/ProcessStep.tsx
@@ -1,10 +1,11 @@
 import { Flex, Heading } from '@/components';
-import { processStep } from '@/constants/processStep';
 import { MAX_RESPONSIVE_WIDTH } from '@/constants/system';
 import { getOpacityColor } from '@/utils/getOpacityColor';
 import { styled, useMediaQuery } from '@mui/material';
 
-const ProcessStep = () => {
+import { processStep } from '../constants/processStep';
+
+export const ProcessStep = () => {
   const isMobile = useMediaQuery(MAX_RESPONSIVE_WIDTH);
 
   return (
@@ -37,8 +38,6 @@ const ProcessStep = () => {
     </>
   );
 };
-
-export default ProcessStep;
 
 const S = {
   StepBox: styled(Flex.Row)<{ isMobile: boolean }>`
