+++ src/pages/DashboardPage/constants/compareOptionLabel.ts
@@ -1,0 +1,6 @@
+export const compareOptionLabel = {
+  term: '동일 학기',
+  entryYear: '동일 학번',
+  major1: '동일 1 전공',
+  major2: '동일 2 전공',
+};
