+++ src/pages/DashboardPage/constants/faq.ts
@@ -1,4 +1,4 @@
 import faqData from '@/data/faqData.json';
-import { FAQItem } from '@/types/faq';
+import { FAQItem } from '../types/faq';
 
 export const FAQ: FAQItem[] = faqData;
