+++ src/pages/DashboardPage/hooks/useGetCapabilityQuery.ts
@@ -1,10 +1,11 @@
-import { getCapability } from '@/apis/capability';
 import { QUERY_KEYS } from '@/constants/queryKeys';
-import { CapabilityResponse } from '@/types/capability';
 import { useSuspenseQuery } from '@tanstack/react-query';
 import { AxiosError } from 'axios';
 
-const useGetCapabilityQuery = () => {
+import { getCapability } from '../apis/capability';
+import { CapabilityResponse } from '../types/capability';
+
+export const useGetCapabilityQuery = () => {
   const { data, ...rest } = useSuspenseQuery<CapabilityResponse[], AxiosError>({
     queryKey: [QUERY_KEYS.capability],
     queryFn: () => getCapability(),
@@ -12,5 +13,3 @@ const useGetCapabilityQuery = () => {
 
   return { capability: data, ...rest };
 };
-
-export default useGetCapabilityQuery;
