+++ src/pages/DashboardPage/hooks/useGetSemesterCapabilityQuery.ts
@@ -1,14 +1,13 @@
-import { getSemesterCapability } from '@/apis/capability';
 import { QUERY_KEYS } from '@/constants/queryKeys';
-import { SemesterCapabilityResponse } from '@/types/capability';
 import { useSuspenseQuery } from '@tanstack/react-query';
 import { AxiosError } from 'axios';
 
-const useGetSemesterCapabilityQuery = () => {
+import { getSemesterCapability } from '../apis/capability';
+import { SemesterCapabilityResponse } from '../types/capability';
+
+export const useGetSemesterCapabilityQuery = () => {
   return useSuspenseQuery<SemesterCapabilityResponse[], AxiosError>({
     queryKey: [QUERY_KEYS.semesterCapability],
     queryFn: () => getSemesterCapability(),
   });
 };
-
-export default useGetSemesterCapabilityQuery;
