+++ src/pages/DashboardPage/types/capability.ts
