+++ src/pages/DashboardPage/types/faq.ts
